Clean up OBO tests.
